+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.1.0
 
 ## [Unreleased]
 
+### Changed
+- Removed `device` parameter from `run_inference()` function - device is now automatically detected for optimal performance
+-  Reorganize the documentation
+
 ## [0.3.2] - 2025-09-15
 
 ### Changed